requirements.txt
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
requests==2.25.1
selenium==3.141.0
urllib3==1.26.2
webdriver-manager==3.2.2

bs4~=0.0.1
beautifulsoup4~=4.10.0
Pillow~=9.2.0